[CHE-408] Refactor existing ApplePayViewModel. Split ApplePayViewModel and ApplePayService functionalities.
